Update deep-q-network.mdx
I am adding a note that we are using Same network Not 2 networks yet. So we know that this will cause issue for training. And then to solve this issue we will use Target Network.
On the next chapter The Deep Q-Learning Algorithm, the picture of pseudocode is showing that we are using 2 networks. It may cause confusion for beginner . Please add some part saying we are using One network first . Thanks !